requirements.txt
langchain_openai
langchain-google-genai
browser_use
python-dotenv
apscheduler
streamlit
requests
gTTS
elevenlabs